backend/requirements.txt
# FastAPI va WebSocket uchun kerakli kutubxonalar
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
redis==5.0.4
